azure_func/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
pillow
pymongo
azure-cognitiveservices-vision-computervision
msrest